Fix for NA handling by CF_detection function
It seems like medepir:::pval_CF does not handle the missing values correctly. It looks like using I(val) in the lm regression formula (entire methylome) instead of each column of val (which corresponds to one CpG) one by one, causes na.omit being applied to the rows (individuals) and not columns (CpGs). Presence of missing values in the methylome may result then in a massive (or total) decrease in the number of individuals tested.